- Created service class - Firebase api is now used only by FirebaseService - ScreenshotForBlocked doesn't know which service it uses
