+++ service.py
@@ -1,0 +1,6 @@
+class Service:
+    def get_last_mention(self):
+        pass
+
+    def set_last_mention(self, last_mention):
+        pass
